CLOUDP-362014: Enable s390x agent test infrastructure with conditional OTel and libsodium support
