+++ requirements.txt
@@ -23,7 +23,9 @@ GitPython==3.1.46
 setuptools>=71.0.3 # not directly required, pinned by Snyk to avoid a vulnerability
 opentelemetry-api==1.40.0
 opentelemetry-sdk==1.40.0
-pytest-opentelemetry==1.1.0
+# grpcio segfaults at process exit on s390x + Python 3.13 and is pulled by pytest-opentelemetry
+pytest-opentelemetry==1.1.0 ; platform_machine != "s390x"
+grpcio==1.80.0 ; platform_machine != "s390x"
 black==26.3.1
 flake8==7.3.0
 isort==8.0.1
