+++ scripts/evergreen/e2e/single_e2e.sh
@@ -109,7 +109,8 @@ deploy_test_app() {
     # As soon as we are having one OTEL expansion it means we want to trace and send everything to our trace provider.
     # otel_parent_id is a special case (hence lower cased) since it is directly coming from evergreen and not via our
     # make switch mechanism. We need the "freshest" parent_id otherwise we are attaching to the wrong parent span.
-    if [[ -n "${otel_parent_id:-}" ]]; then
+    # pytest-opentelemetry and grpcio are not available on s390x, so OTEL tracing is disabled there.
+    if [[ -n "${otel_parent_id:-}" ]] && [[ "${PYTEST_OTEL_DISABLED:-false}" != "true" ]]; then
         otel_resource_attributes="evergreen.version.id=${VERSION_ID:-},evergreen.version.requester=${requester:-},mck.git_branch=${branch_name:-},evergreen.version.pr_num=${github_pr_number:-},mck.git_commit=${github_commit:-},mck.revision=${revision:-},is_patch=${IS_PATCH},evergreen.task.name=${TASK_NAME},evergreen.task.execution=${EXECUTION},evergreen.build.id=${BUILD_ID},evergreen.build.name=${BUILD_VARIANT},evergreen.task.id=${task_id},evergreen.project.id=${project_identifier:-}"
         # shellcheck disable=SC2001
         escaped_otel_resource_attributes=$(echo "${otel_resource_attributes}" | sed 's/,/\\,/g')
